migrate webhook IDs to 64-bit integers (#2444)
Fixes #2443 (other fixes needed)

Related to the repo ID change made by Evan recently: use 64-bit ints for
webhook IDs. This involves increasing the size of integer we use in the
DB record, and making logic changes where needed.

+++ internal/controlplane/handlers_githubwebhooks.go
@@ -370,7 +370,7 @@ func (s *Server) deleteWebhookFromRepository(
 
 	webhookId := dbrepo.WebhookID
 	if webhookId.Valid {
-		resp, err := client.DeleteHook(ctx, dbrepo.RepoOwner, dbrepo.RepoName, int64(webhookId.Int32))
+		resp, err := client.DeleteHook(ctx, dbrepo.RepoOwner, dbrepo.RepoName, webhookId.Int64)
 		if err != nil {
 			if resp != nil && resp.StatusCode == http.StatusNotFound {
 				// if the hook is not found, we can ignore the error, user might have deleted it manually
